Add Icinga data caching support (local-only)
All requests to Icinga are now routed through [groupcache](https://github.com/mailgun/groupcache)
before being sent to the Icinga server, which means that duplicate requests for
the same data will not hammer Icinga if they're made in quick succession.

While groupcache has the ability to share its results amongst multiple Meerkat
servers, for now that functionality is not enabled, while we assess the
stability and performance implications of the current approach.