Remove docker-compose workflow
From the first line of Docker's docker-compose documentation:
"Compose is a tool for defining and running multi-container Docker applications"
https://docs.docker.com/compose

Meerkat is a single-process application. It could only ever require one container lol.

Removing docker-compose simplifies the documentation (this change updates the README)
and is one less workflow to maintain.
The development workflow could be made easier still, but those are separate changes.