backend: require authentication to edit dashboards if set
This is a dumb implementation that is really the bare minimum. You can
set one username password pair in the configuration file, then any
attempt to edit a dashboard must first go through HTTP Basic
Authentication with those credentials. No support for more than one
set of credentials. I'm not entirely convinced that this application
needs to be run on a server connected to the network anyway.

A later change could render the homepage differently based on whether
it has successfully authenticated. For example, render on the "View"
and not "Edit" or "Delete" dashboard buttons.

See #31