frontend: filter all src through prettier
This needed to be done at some point. This is a gross diff but this is
one of those "rip off the bandaid quickly" type scenarios. Now both
the frontend and backend code formatting and style is automated.
This hooks in to any new (VSCode) and old (vi) editors I can think of.

I don't particularly like the output of prettier, but gofmt is
nobody's favourite either. That's the point!