Test and build frontend with Github Actions
We're not really testing; just lint and code formatting.
But hey, gotta start somewhere!
See #57

I'm not sure if this is an appropriate way to use github actions.
I don't know if I should create a different workflow or keep adding steps to the workflow.
Need to do more investigation and see what other projects on Github are doing.
But for now, to get a nice green tick, let's see how this works...