frontend, backend: only show edit, delete dashboard when authenticated
This works by first testing if authentication is configured on the server.
If so, authenticating to the server results in a cookie being returned.
The presence of a cookie changes how the dashboard list is rendered.

Cookies are generated by the server by creating "tokens": base64-encoded random bytes.
Tokens are only held in-process in memory.

I don't know how to create a "log in" button to prompt authentication
A <button> element which has its onClick attribute set to
meerkat.authenticate() may do the trick.

Closes #96