Add script for creating a meerkat tarball
Creates a distribution tarball in a temporary directory from source.
The path to the created tarball is the last line of the script's
output.

This is part of the way to getting to a debian package or container
image etc.. We can publish the tarball as a thing to download in a
github release thing.