frontend: reliably justify "check-content" text
The flex container was only applied to direct text child elements of the check class:

	.check > .text { ... }

It's clearer to apply the flex property directly to the check-content
class we were already using. This way we're able to eliminate the
source of the bug in
https://github.com/meerkat-dashboard/meerkat/issues/115.
Unclear inheritance rules resulted in unintuitive behaviour. Tested ok
with static text, static ticker, icinga card elements.

Resolves https://github.com/meerkat-dashboard/meerkat/issues/115