frontend: update bootswatch v5.2
This is an upgrade from v4.2 which was released in 2018! May as well
get a more "modern" appearance for free!

This also swaps out the minified CSS for regular CSS. Saving 30K is
not really important when the single stylesheet is loaded from cache
>90% of the time. Also Meerkat is designed for extremely low traffic
on a trusted LAN running on top of (at least!) a gigabit ethernet. No
point sacrificing readability, discoverability for marginal benefit.
False economy! More significant slimming down could be done by dumping
dependencies.

Updated via

	curl https://bootswatch.com/5/darkly/bootstrap.css > frontend/assets/bootstrap.css

We should keep this as a shell script so that it's clear how this
program is put together, and for future updates!