frontend: Use form class styles from bootstrap
Instead of specifying, and overriding, core HTML element styling we
can let bootstrap do all the fiddle-y stuff for us. They've worked out
all the painful stuff of alignment so if we're pulling the dependency,
we should lean on it when we can. Otherwise we're left with our own
inconsistent styling bugs!!