all: Always name the app Meerkat
Anyone who wants to distribute Meerkat with a different name are free
to do so under the open source licence Meerkat is published under.
Given the amount of code, and how it required both frontend *and*
backend cooperation to to implement this feature, it's clearer for
everyone to always call this app "Meerkat".