frontend: use the idiomatic left-right header
This is what is used in lots of web apps so rather than rolling our
own, we can use this style so we get a feel the app without having
really used it before. This means we can "borrow" logic from other
projects ;)

While here, remove our a CSS selector. This is managed by bootstrap;
we don't have to do this ourselves