frontend: extract font size input to its own component
The same thing copied and pasted around 5 times. These components are
pretty hefty so it's nice to do whatever we can to get the size of
them down!

This patch addresses stuff brought up in
https://github.com/meerkat-dashboard/meerkat/issues/104 Specifically -
the more we factor out of the elements the easier it is to make their
behaviour consistent where appropriate.