frontend: tidy element URL render
No need for custom classes; we can just set the href to the URL. Plus
some trivial logic changes to shorten the code significantly. We also
don't need to build up a HTTP URL if just given a hostname. The form
says URL, and a hostname is not a URL, so there's a possibility of
surprise there, especially when this "feature" was not documented or
presented to the user!