all: (temporarily?) remove dashboard templating feature
The implementation of this variables feature was hairy: it's
undocumented and not clearly specified in the code.

A suggested way forward is writing tiny programs (scripts!) which
create and dashboards serialised as JSON. A python (or whatever!)
script is a clearer expression of creating dashboards than globally
available key-value pairs sometimes read by some react components,
spread across different files with differing rules on when and how to
rerender.

Even in Grafana which has millions of investor dollars behind
development, dashboard variables are hard to get your head around and
lead to confusing dashboard state!!