frontend: Clear all options on element type change
This got confusing. Elements could accrue all sorts of irrelevant
values in their options which did not effect how they behaved or
looked.

This patch also extracts the default element options logic to its own
function. The ElementSettings component is really big. This makes it a
bit smaller to make it easier to see what is going on.

While here use consistent option name for font size with all elements.