backend: don't track names of options we don't care about
The backend doesn't care, and can't really do anything with, all the
possible options that the frontend may use. In this part of the
application, the backend is really just a place to store a serialised
version ot elements; no reason to marshal it into a type when we don't
have any code to deal with the type at all.