all: embed web app into server, let server route all requests
This change isn't really as big as it looks. The interactive parts of
meerkat are handled by preact, and the rest is done by the server. The
server serves basic HTML, CSS, and a path to the relevant script which
renders the UI (home, viewer or editor).

Webpack now manages 3 outputs for each primary "view". It's a
multi-page application.

From there, we could drop the preact-router dependency (and the 4
security vulnerability warnings we kept getting!).

The rest of the work was in swapping out the state management done by
passing around URL parameters to the standard react diffing algo stuff
used by the rest of the UI. For simplicity in state management, an
element must be selected in the UI for it to be transformed. To make
selection easier, any part of an element can be clicked to select it
(instead of a little "Edit" button).