all: serve static content direct from server
These parts of the meerkat UI do not need to be rendered by a complex
javascript engine. It is relatively static content which has little to
no interactivity requirements. Using server side rendered HTML for
these parts of the interface is easier to reason about and is easier
to maintain in the long run. No need to understand complicated
frontend frameworks: just read the HTML files to see what is going on.
The Go standard library is significantly easier for maintenance than
these ever-changing javascript dependencies.