main: let page load before layout
On every single page load, we would get this warning printed to the
browser console:

"Layout was forced before the page was fully loaded. If stylesheets are
not yet loaded this may cause a flash of unstyled content."

Loading the preact script in the head of the HTML is strongly
discouraged I discovered. As a bonus we don't need to render a
different head, we can use the template default.

While here suppress useless loading/no dashboard message