frontend: Keep object state in with the component
The only place we get object state is by polling, so we can keep all
that in the component until we implement something more clever. This
breaks the regular expression matching on the check result output text
for now; we'll come back to it when the rest of this module is tidied up.