all: Allow serving the UI from the local OS filesystem  too
This splits up a lot of functionality from the main package. The main
package doesn't need to know (or care) how the user interface is
served. It just cares about net/http Handlers. Now that there's more
than one package we can start isolating and documenting different
parts.

There's a lot of functionality which is easier to understand when it is separated from the main package.