ui: make audio/video support generic
The browser and HTML5 transparently supports "streams" or files for
us. We don't have to come up with our own names and re-implement the
difference between a stream and a file when the interface exposed by
the browser doesn't even let us specify that difference!

This patch removes all references to HLS stream, iframe-video,
Iframe... There are no iframes. All of this stuff has been provided by
the HTML video element since 2009.