cmd/meerkat, ui: Load images by URL only
This is how the web is designed! Instead of needing to download then
upload a file to meerkat, we can link to the file. The browser loads
the file as normal and we're done. This reduces the state in meerkat.
It's easier to expose to the network/internet, since nobody can upload
arbitrary files any more. Upgrade/migration documentation is included.

Closes https://github.com/meerkat-dashboard/meerkat/issues/119
Closes https://github.com/meerkat-dashboard/meerkat/issues/93
Closes https://github.com/meerkat-dashboard/meerkat/issues/130