all: Create a dashboard info page
The dashboard info page contains information and controls for a
specific dashboard. For the relatively rare operations of changing
dashboard title and background, we can move basic CRUD operations out
of the complex editor app and into boring HTML forms in the info page.
With an information page, we can put other more rarely used controls,
like Delete, out of the home page. With the home page looking quite
bare - just a list of dashboards - using a JS framework for
interactive pages seemed like overkill, so this patch removes all
javascript from the homepage.

The only interactivity provided by the javascript was the "filter
bar". But that functionality has already been provided by the browser
for at least 20 years (ctrl-f or command-f).