ui: remove unused static-ticker element
From a sample of over 300 dashboards, 2 dashboards used a single
static ticker element.

One was a test dashboard. The other was a dashboard which displayed an
out-of-date message "Meerkat maintenance this afternoon" which could
be displayed fine and more clearly with text which doesn't move!

We've been here before; the <marquee> has been and gone

The code to make this work was not good. Special conditions all over
the place just for scrolling text.