ui: remove unused check image element
In a sample of approx 300 dashboards, this element was not used
anywhere. It was (ab)used as a navigation button in twelve dashboards,
which when clicked would take you to another dashboard. A side effect
of this is use was invalid requests were being sent to the server,
looking for an icinga object of type "undefined". This patch will
eliminate some of the many invalid requests seen in real environments.

Includes updates to documentation