ui: use CSS to consistently scale lines, arrowheads
Letting the browser handle scaling means we avoid the circular
dependency using a ref on the DOM node to work out how big the line
should be - before we the node was rendered! This also replaces the
custom 'arrowhead' implementation with SVG markers intended for this
purpose.

Gets rid of a superflous div element.

Resolves #150