ui: allow rendering any icinga object attribute from a collection
Previously it was only possible in icinga cards to select general
attributes like name and state calculated from a collection of objects
(such as from a group or result of a filter expression). This change
enumerates the attributes from every object of a collection and allows
selecting any of them. This isn't particularly well done as the
selector for an attribute is something like
"source.1.attrs.last_check_result.output" for the output of the check
command of the 2nd object in a list (index 1 counted from 0). The
index is used as our selectors expressions cannot escape ".".

Resolves #148