ui: Select icinga obj attrs by obj name
Object attributes are now keyed by object name rather than array
index. This makes it possible to eyeball the attributes to get an idea
of what to select. From there, we can make the list an autocompleting
text field. Instead of scrolling through, say, 100+ options, we can
type some text to filter attributes we may be interested in.

Resolves #162